+++ packages/dart_client/lib/src/api/channel.dart
@@ -428,6 +428,8 @@ class Channel {
   /// Mark all channel messages as read
   Future<EmptyResponse> markRead() async {
     _checkInitialized();
+    client.state.totalUnreadCount =
+        max(0, (client.state.totalUnreadCount ?? 0) - (state.unreadCount ?? 0));
     state._unreadCountController.add(0);
     final response = await _client.post('$_channelURL/read', data: {});
     return _client.decode(response.data, EmptyResponse.fromJson);